fix: 将订单 (orders) 领域中涉及金额和货币的部分从 BigDecimal 类型更改为 long 类型, 并将货币换算等逻辑收敛到 CurrencyConfig 中

+++ international-shopping-infrastructure/src/main/java/shopping/international/infrastructure/dao/orders/po/CartItemViewPO.java
@@ -2,7 +2,6 @@ package shopping.international.infrastructure.dao.orders.po;
 
 import lombok.*;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -55,6 +54,5 @@ public class CartItemViewPO {
     /**
      * 展示单价
      */
-    private BigDecimal unitPrice;
+    private Long unitPrice;
 }
-
