+++ international-shopping-infrastructure/src/main/java/shopping/international/infrastructure/dao/orders/po/OrderDiscountAppliedPO.java
@@ -3,7 +3,6 @@ package shopping.international.infrastructure.dao.orders.po;
 import com.baomidou.mybatisplus.annotation.*;
 import lombok.*;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -46,11 +45,10 @@ public class OrderDiscountAppliedPO {
      * 实际抵扣金额
      */
     @TableField("applied_amount")
-    private BigDecimal appliedAmount;
+    private Long appliedAmount;
     /**
      * 创建时间
      */
     @TableField(value = "created_at", insertStrategy = FieldStrategy.NEVER, updateStrategy = FieldStrategy.NEVER)
     private LocalDateTime createdAt;
 }
-
